Remove video and other homepage elements: #11
Remove intro video from top of page. It is
still available from the WWT YouTube channel (linked below).
Remove secondary menu across the top of the carousel,
as discussed with website committee.

Remove Twitter and News sections, since they are largely
obsolete.  For now, this means that we aren't linking to the
blog posts that had been linked from the News section.  Maybe
we should add a "Press" or "Media" section and link to those
posts from there.  (Or perhaps that already exists somewhere?)