Add initial RequestProvider infrastructure
This will allow us to move code out from the code behind. Once that is done, we can more easily refactor things like file access so we can replace it with Azure access. This PR puts in the basics to enable this kind of workflow while showing an example for the 2MASSOct.aspx endpoint.

The goal is for the aspx pages to essentially contain nothing besides:

RequestProvider.Get<Name>().Run(Request, Response)

The providers are currently just manually new'd up, but can be replaced with a DI (or for now probably ServiceLocator) pattern once that's in. The base class wraps the request and response objects in a context so derived types aren't depending on System.Web (they are indirectly for now, but this helps make it easier to redirect stuff in the future).