Convert all WWTWeb endpoints to use RequestProvider
This change moves most code behind to a class library. This will help
isolate out code so that tooling can be used to simplify refactoring
things such as data access. It also simplifies code so that tooling can
show where methods are being used (code behind files such as aspx and
aspx.cs do not work well for that).