Convert PlateUploader to more generic PlateManager cli
This removes the dependency on WinForms, and provides a single tool that can be used to manage plate file tasks. A pub/sub model is used via the System.Threading.Channels namespace. This allows for a pipeline infrastructure to be used to process the files into images and then to upload them.