Add tests for DSS provider
This required a bit of refactoring of the WwtContext so that it could be testable. Instead of being a struct, it is now an interface that can accept any implementation. This is now mockable from within the testing framework. This also abstracts out the configuration into an options type that defaults to ConfigurationManager, but allows for cusomizations if needed.