Remove WWTUtil.GetCurrentConfigShare (#116)
This whole infrastructure implements a basic failover scheme that isn't
even used in the current production system, and it certainly isn't
something that we want to bring back in the new system.

This change introduces a small behavior change because
GetCurrentConfigShare called `ToLower()` on the config value that it
returned. We don't do that here, which shouldn't cause any problems.