Replace abstractions with interfaces
A while back, we swapped out a struct implementaiton of WwtContext with
an interface. At that time, Http*Base types were used as abstractions
over the request/response objects. These are possible to mock, but not
as easy as interfaces. This change swaps these out for pure interfaces
so that mocking is much easier and as a plus removes some dependency on
System.Web.

An implementation of this for Page is still used for the aspx pages.
However, that implementation implements all the interfaces (explicity,
ie privately) so that only a single object must be constructed to pass
the context around.