Use Microsoft.Extensions.DependencyInjection instead of Unity for requests (#144)
We're currently using unity for dependency injection and aren't using any features that are specific there. MS.DI is a simpler abstraction and has some simpler ways of doing things like decorators so that we can easily do things such as adding a caching layer on top. This will also make it easier to move to ASP.NET Core when we're ready to do that.

The containers for the community site had already been its own set of registrations, so this reverts back to that pattern so the communities site continues to use Unity.