Handle case sensitivty (#162)
This changes does the following to handle case sensitivty:

- Ensure that the ids for tours are lower case

Tour ids are GUIDs and are already lower cased in storage. This change
ensures that the id passed in lower cased so it will match the storage
entry.

- Add IKnownPlateFiles to limit files exposed by tiles[2].aspx

A file has been placed in `coredata/known_plate_files.txt` that is used
for the known plate files. It is a simple list where each line is a
plate file name. On the current VM, there are 224 files that are exposed
by these endpoints. This class helps ensure only those are exposed, as
well as normalize their casing to what is in the blob.

- Lower cases the thumbnail name and class.

Thumbnails appear to be addressed by inconsistent casing. All thumbnails
will be lowercased in storage. This change lowercases the name that is
searched so it will match once the data is updated.