Revert to old martiantile.aspx with updates
This brings back the logic from the previous martiantile.aspx, but
updates it to use the Azure access patterns.

However, it accesses a platefile via exploded png values, so a layer was
added to the plate tile pyramid accees level to know how to handle that.
I think in the long term, we may want to change this to more of a
pipeline where a chain of IPlateTilePyramid implementations can identify
which platefiles they know how to handle; for now, inheriting is solving
the problem.