Add .NET 5 project head
This adds initial support for ASP.NET Core running on .NET 5. Thanks to
the previous refactoring work, there is very little that needed to be
done to support this.

With this change, the wwturl tests all pass against .NET 5. There may be
some issues not captured by this, but looks good for an initial push.