Attempt to the server's OAuth login bits working again
The OAuth flow stopped working, seemingly because our redirect_uris are
now being rejected. This change may not entirely solve the problem, but
hopefully it will get the server's OAuth bits to work. We replace WWT
redirect URIs with the special "desktop" one, which is actually more
correct since the server-side OAuth flow is app-like rather than running
in a web browser.