Update to .NET 5 wave of libraries
Now that we're moving towards support .NET 5, this will update all usage of the 3.1 wave to 5. The .NET 5 app would already be using the new libraries, so now the WWTMVC5 app will as well.

This also updates the Redis library which should help with the exceptions we've been seeing under load.