Add initial disabled nullable directive (#232)
Nullable reference types are a way of annotating types that can or cannot be null. It's a new feature added in C# 8.0 that can drastically limit the number of null reference exceptions in a running app. Since many of the issues in deployment have been around uncaught nulls, it makes sense to set the project up to enable them.

It's quite daunting to add them to a code base at one time, so the general pattern is to turn it on globally, then turn them off for each page. This change accomplishes that. Future changes can then handle enabling them on a class-by-class basis. An effect of this change is that any new classes will be forced to handle nullable reference types, which limits new code from introducing null reference exceptions.