Attempt to fix OAuth redirection with deployment slots
It turns out that if we're going to have "stage" and "production"
deployment slots that we're going to hot-swap, they need to have the
same ASP.Net app settings. (This makes sense in retrospect since if they
had different settings, you'd have to change the settings on-the-fly
during the swap, which isn't possible.) So the method that we had for
setting the OAuth redirection URLs isn't tenable. Instead, we need to
configure a dictionary of potential URLs and choose one at runtime based
on the HTTP Host header.