WWTMVC5: take 2 of handling old content URLs
Based on my research it seems that URL paths with filename extensions
don't even get passed to the server code in most cases, or at least not
in a way where can intercept some requests but not other requests for
static files.

For take 2, try hardcoding some manual rewrite rules in the web.config.
It is dumb and bad but it seems like the most tractable option.