WWTMVC5: work around new spurious CS1647 error
As recently, the ASP.NET build-views stage hits a CSS1647 "expression is
too long or complex to compile" error, which seems to be due to it
trying to compile `wwtsdk.aspx` even though it's a Content item. I can't
find a way to get the compiler to ignore it, but there is a way to just
delete it so that the compiler doesn't see it. This shouldn't affect the
actual file service.

Note that, based on separate tests, this is unrelated to the .NET 6
update. The only difference from previous successful builds that I can
see is that the Pipelines "Visual Studio build" task version is now at
1.208.0. Builds worked with the previous 1.1xx version.