src/WWT.Providers/TileProviders/NLASDSSProvider.cs: add this
A new provider to serve up a blend of the NOIRLab All-Sky map and WWT's
classic DSS. Level 0 is pure NLAS, levels 9-12 are pure WWT DSS, and in
between we've computed a blend between the two. There's inevitably some
unevenness in the intermediate levels, but when used as context for
high-resolution imagery that should be less noticeable.

This relies on a new `nlas_dss_rev1.plate` plate file uploaded to the
`coredata` blob storage container.