Start process to move endpoints to minimal apis
During the migration to ASP.NET Core, the APIs were moved from WebForms aspx pages into a shared format that could be consumed by ASP.NET Framework and Core. Now that we're targeting .NET 8 only, it would be ideal to start moving away from the intermediary form (i.e. RequestProviders) and use minimal APIs directly.

This change starts that with the following to show the pattern:

- /wwtweb/dss.aspx - the request provider is turned into a web-agnostic service that just returns a stream
- /wwtweb/mandel.aspx - remove unnecessary service and uses IMandelbrot directly
- deprecated endpoints - this was being done with a request provider but can be handled easily by a minimal api directly

This also removes the test for dss.aspx that doesn't seem to be doing much beside a copy of the implementation.